git-svn-id: svn://141.76.65.164/kp-salespoint@159 eb53f5e8-9ec5-4bc0-8d13-06c033283886

+++ SalespointFramework/src/org/salespointframework/core/calendar/AbstractCalendar.java
@@ -84,7 +84,7 @@ public abstract class AbstractCalendar<T extends CalendarEntry> implements Calen
     }
 
     @Override
-    public void deleteEntry(int calendarEntryIdentifier) {
+    public void deleteEntry(long calendarEntryIdentifier) {
         em.remove(this.getEntryByID(calendarEntryIdentifier));
     }
 }
